changed createmenu to setpossiblecommants. added code to get the possible commands. when we use ncurses, the menu will print. when we dont't, the command will print.
changed the commandparser. when ncurses is used, it will get the line from the menu. when ncurses is not used, it will get it from cin.

added showpossiblecommands() to startclass.cpp. this presents a little bug that needs to be fixed. when the game is started for the first time, for the first command must be clicked twice.

in room.cpp, deleted the \r from \r\n